Update MySensors.MySGW return for state map
`init` returned `port` for the state, but in the terminate
it is attempting to use `state.port` if it existed.

Instead of changing terminate to just look for the port as
the state, this updates the setup of the state to not just
be the port, as it allows it to be more extensible in the
future.